Add change of language logic
More or less temporary implementation. Complete dynamic is not possible
with the current built-in i18n. The solution for the moment is
to change the state and reload. Language is therefore loaded
as the page loads.

Changing of language introduced in the info area. Should be moved to
something like program settings eventually.